Refactor EPUB processing logic and add parser module
